backend/requirements.txt
azure-functions
azure-storage-blob
azure-storage-queue
requests

